Fix multilayer with fields
Decoder expects that 2 fields of one frame are stored in pcListPic next to each other.

But in case of Multilayer there will be some distance between field pictures. For example:

- POC0 (TopField) Layer0
- POC0 (TopField) Layer1
- POC0 (TopField) Layer2
- POC1 (BottomField) Layer0
- POC1 (BottomField) Layer1
- POC1 (BottomField) Layer2

Decoder needs to search for a second field picture.
It should belong to the same layer as the first field picture.

Also, POCs of first and second field are not necessary differ by one. So, I removed those checks