Fix parsing of unknown SEI message type.
When decoder parses SEI message of unknown type,
it does not add it to list of SEIs.

But then try to access it by calling back() of the list of SEIs.

If list is empty decoder crashes.

Otherwise (if list is not empty) decoder gets wrong SEI message (one of previously added to the list)

Possible fix is to make xReadSEImessage() and parseSEImessage() report if new SEI message object was added to the list