Avoid seg fault Slice::constructRefPicList
When decoding a subset of temporal layers, it is possible that
inactive reference pictures may not be present in the DPB. In
such case a segmentation fault could be observed in
Slice::constructRefPicList. Marking of a picture as long term or
short term is changed to be skipped when a picture is missing.
An error is thrown when an active reference picture is missing.