Initialize m_numSubPic for unavailable pictures.
Set m_numSubPic to a value signaled in SPS for the layer.
Otherwise an error is triggered when an unavailable picture
happens to be in the reference list for GDR picture partitioned into
more than 1 sub-pictures.

Error message is: "TMVP cannot be enabled in pictures
that have no reference pictures with the same resolution"